feat(cli): adds support for params to starter cli
User can now generate starter kits providing --name and --kit params. If a parameter is not used it will start in interactive mode as normal to prompt for any missing args
