+++ starters/angular-ngrx-scss/README.md
@@ -65,13 +65,13 @@ The `state` directory is where all of our NgRx logic lives. You'll find a folder
 ### CLI (Recommended)
 
 ```bash
-npx @this-dot/create-starter
+npm create @this-dot/create-starter --kit angular-ngrx-scss
 ```
 
 or
 
 ```bash
-yarn create @this-dot/starter
+yarn create @this-dot/starter --kit angular-ngrx-scss
 ```
 
 - Follow the prompts to select the `angular-ngrx-scss` starter kit and name your new project.
